Create PlayerCam
This script goes with 1st person camera setup.  You must create Empty object and name is CameraHolder. Then placed the camera "PlayerCamera" inside the empty.  Add the script onto the PlayerCamera, Within the settings. SensX = 800, SensY = 800.  Orientation = plugged the PlayerOrientation (transform) (PlayerOrientation is an empty object that is a child under "Player" group object.  This goes with the playerObj and Camera PoS)